requirements.txt
Flask
Flask-RESTX
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
werkzeug==2.3.3
